requirements.txt
attr==0.3.2
ConfigParser==7.2.0
cryptography==46.0.7
dl==0.1.0
docutils==0.16
fastapi==0.135.3
filelock==3.25.2
HTMLParser==0.0.2
importlib_resources==7.1.0
inflect==7.5.0
ini2toml==0.15
ipython==8.12.3
ipywidgets==8.1.8
jnius==1.1.0
joblib==1.5.2
keyring==25.7.0
matplotlib==3.8.0
google-cloud-storage==2.9.0
pandas==3.0.2
path.py==12.5.0
protobuf==7.34.1
pydantic==2.13.0
pyOpenSSL==26.0.0
pytest==9.0.3
python-dotenv==1.2.2
redis==7.4.0
scikit_learn==1.8.0
seaborn==0.12.2
Sphinx==9.1.0
streamlit==1.28.0
SQLAlchemy==2.0.49
requests==2.31.0
thread==2.0.6
trove_classifiers==2026.1.14.14
typing_extensions==4.15.0
urllib3_secure_extra==0.1.0
uvicorn==0.34.0
xmlrpclib==1.0.1
xx==3.3.2
wordcloud==1.8.2
